requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
pyyaml>=6.0.1
chromadb
requests
pypdf